requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.45.0
blinker==1.7.0
bottle==0.13.2
bottle-websocket==0.2.9
breadability==0.1.20
certifi==2024.6.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
dlib==19.24.4
docopt==0.6.2
Eel==0.18.1
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==25.1.0
filelock==3.18.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
fsspec==2025.3.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
huggingface-hub==0.29.3
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.3
lxml==5.3.1
MarkupSafe==2.1.5
matplotlib==3.10.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.5
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.1
pefile==2023.2.7
pillow==10.3.0
proglog==0.1.10
pyasn1==0.6.1
pycountry==24.6.1
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMySQL==1.1.1
pyparsing==3.2.0
pyperclip==1.9.0
PySimpleGUI==5.0.7
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.40
sumy==0.11.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.2.12
tokenizers==0.21.1
torch==2.6.0
tqdm==4.66.4
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.2
wxPython==4.2.2
zope.event==5.0
zope.interface==7.2
